Move server.js into app folder
With index.js now at the root it makes sense to us to move server.js into the `app/` folder with the rest of the application code.
